+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # avro-builder changelog
 
+## Unreleased
+- Add support for field `order` attribute in serialized output.
+
 ## v2.1.0
 - Add support for Avro 1.12.